app/requirements.txt
flask==0.12
Flask-Caching==1.2.0
Flask-Compress==1.3.0
flask-sqlalchemy==2.2
sqlalchemy-migrate==0.11.0
mysql-connector-python-rf==2.2.2